Feat/formatter

+++ backend/tests/setup.ts
@@ -4,21 +4,20 @@ import fs from 'fs';
 import { v4 as uuidv4 } from 'uuid';
 
 export async function createTestDb(): Promise<Database.Database> {
-	const migrationFile = `./tests/.migrations-${uuidv4()}`;
-	const testDb = new Database(':memory:');
-	testDb.pragma('journal_mode = WAL');
+  const migrationFile = `./tests/.migrations-${uuidv4()}`;
+  const testDb = new Database(':memory:');
+  testDb.pragma('journal_mode = WAL');
 
-	const testUmzug = createUmzug(testDb, migrationFile);
-	await testUmzug.up();
-	(testDb as any)._migrationFile = migrationFile;
+  const testUmzug = createUmzug(testDb, migrationFile);
+  await testUmzug.up();
+  (testDb as any)._migrationFile = migrationFile;
 
-	return testDb;
+  return testDb;
 }
 
 export function cleanupTestDb(db: Database.Database) {
-	const migrationFile = (db as any)._migrationFile;
-	if (fs.existsSync(migrationFile))
-		fs.unlinkSync(migrationFile);
+  const migrationFile = (db as any)._migrationFile;
+  if (fs.existsSync(migrationFile)) fs.unlinkSync(migrationFile);
 
-	db.close();
+  db.close();
 }
